[hotfix][table, tests] Increase managed memory to 200m in BatchAbstractTestBase
By design of FLIP-53, sharing slot will allocate the physical slot according to the slot sharing group’s overall resources.
However, with such a change, some table batch tests would hang in slot allocation since they have a pipelined region(corresponding to a slot sharing group) that have managed memory sum to be more than that of a slot.
Thus we need to increase the taskmanager managed memory. 200m is a proper value which is also verified to work.